Change orientation of table to maintain page width
The existing table layout was too wide. It was not rendered properly
unless the man page was viewed in full screen. This layout will keep it
within the standard character width.

Add a macro to call the `tbl` preprocessor '\" t

Correct an error in the table for the "excercise" mode (missing
newline).